% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{attr_style}
\alias{attr_style}
\title{Create the content of the html style attribute}
\usage{
attr_style(...)
}
\arguments{
\item{...}{named arguments are used as settings in the html style attribute, with the name being the name of the setting (e.g., background-color). All arguments must be vectors of the same length. NA values can be used to ignore a setting, and if all settings are NA then NA is returned (instead of an empty string for style settings).}
}
\value{
a character vector with the content of the html style attribute
}
\description{
Designed to be used together with the html_attributes function.
}
