% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.r
\name{close_html}
\alias{close_html}
\title{close html file}
\usage{
close_html(template, filename)
}
\arguments{
\item{template}{The html header/footer template}

\item{filename}{The name of the file to save the html. Default is a temp file}
}
\description{
see documentation for \link{create_html}
}

