% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenvis.r
\name{colorscaled_reader}
\alias{colorscaled_reader}
\title{Convert tokens into full texts in an HTML file with color ramp highlighting}
\usage{
colorscaled_reader(tokens, value, alpha = 0.4, meta = NULL,
  col_range = c("red", "blue"), doc_col = "doc_id",
  token_col = "token", filename = NULL, ...)
}
\arguments{
\item{tokens}{A data.frame with a column for document ids (doc_col)
and a column for tokens (token_col)}

\item{value}{A numeric vector with values between -1 and 1. Determines the color
mixture of the scale colors specified in col_range}

\item{alpha}{Optionally, the alpha (transparency) can be specified, with 0 being fully
transparent and 1 being fully colored. This can be a vector to specify a
different alpha for each value.}

\item{meta}{A data.frame with a column for document_ids (doc_col). All other columns are added
to the reader as document meta}

\item{col_range}{The color used to highlight}

\item{doc_col}{The name of the document id column}

\item{token_col}{The name of the token column}

\item{filename}{Name of the output file. Default is temp file}

\item{...}{Additional formatting arguments passed to create_reader()}
}
\value{
The name of the file where the reader is saved. Can be opened conveniently from within R using browseUrl()
}
\description{
Convert tokens into full texts in an HTML file with color ramp highlighting
}
