% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.r
\name{create_html}
\alias{create_html}
\title{create html file}
\usage{
create_html(template, filename = NULL)
}
\arguments{
\item{template}{The html template file, as created with \link{html_template}}

\item{filename}{Name of the file in which to save the html}
}
\value{
the filename
}
\description{
For creating the html in batches, the create_html(), fill_html() and close_html() functions can be used.
create_html() and close_html() require the template for creating the header and footer.
fill_html() can be used sequentally to fill the html.
}
\details{
The connection closes between each fill, to prevent incidental writing to the file.
}
