% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_misc.r
\name{create_meta_tables}
\alias{create_meta_tables}
\title{HTML tables for meta data per document}
\usage{
create_meta_tables(meta, ignore_col = NULL)
}
\arguments{
\item{meta}{a data.frame where each row represents the meta data for a document}

\item{ignore_col}{optionally, a character vector with names of metadata columns to ignore}
}
\value{
a character vector where each value contains a string for an html table.
}
\description{
Each row of the data.frame is transformed into a html table with two columns: name and value.
The columnnames of meta are used as names.
}
\examples{
tabs = create_meta_tables(sotu_data$meta)
tabs[1]
}
