% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.r
\name{fill_html}
\alias{fill_html}
\title{Fill html file}
\usage{
fill_html(data, filename)
}
\arguments{
\item{data}{The html body data}

\item{filename}{The name of the file to save the html. Default is a temp file}
}
\description{
see documentation for \link{create_html}
}
