% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenvis.r
\name{highlighted_reader}
\alias{highlighted_reader}
\title{Convert tokens into full texts in an HTML file with highlighted tokens}
\usage{
highlighted_reader(tokens, value, meta = NULL, col = "yellow",
  doc_col = "doc_id", token_col = "token", filename = NULL, ...)
}
\arguments{
\item{tokens}{A data.frame with a column for document ids (doc_col)
and a column for tokens (token_col)}

\item{value}{Either a logical vector or a numeric vector with
values between 0 and 1. If a logical vector is used, then tokens
with TRUE will be highlighted (with the color specified in pos_col).
If a numeric vector is used, the value determines the alpha (transparency),
with 0 being fully transparent and 1 being fully colored.}

\item{meta}{A data.frame with a column for document_ids (doc_col). All other columns are added
to the reader as document meta}

\item{col}{The color used to highlight}

\item{doc_col}{The name of the document id column}

\item{token_col}{The name of the token column}

\item{filename}{Name of the output file. Default is temp file}

\item{...}{Additional formatting arguments passed to create_reader()}
}
\value{
The name of the file where the reader is saved. Can be opened conveniently from within R using browseUrl()
}
\description{
Convert tokens into full texts in an HTML file with highlighted tokens
}

