% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.r
\name{html_template}
\alias{html_template}
\title{create the html template}
\usage{
html_template(template, css_str = NULL, col1 = "#7D1935", col2 = "#F5F3EE")
}
\arguments{
\item{template}{The name of the template to be used}

\item{css_str}{A character string, to be directly added to the css style header}

\item{col1}{The first style color (top bar color)}

\item{col2}{The second style color (background color)}
}
\value{
A list with the html header and footer
}
\description{
create the html template
}
