% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lda_mod}
\alias{lda_mod}
\title{Word assignments, docXtopic matrix and topicXword matrix of an LDA model of the SOTU data}
\format{lda_mod: Word assignments is a data.frame with document, lemma and topic columns. topic_word_mat and doc_topic_mat are matrices}
\usage{
data(lda_mod)
}
\description{
Word assignments, docXtopic matrix and topicXword matrix of an LDA model of the SOTU data
}
\keyword{datasets}
