% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.r
\name{save_html}
\alias{save_html}
\title{Wrap html body in the template and save}
\usage{
save_html(data, template, filename = NULL)
}
\arguments{
\item{data}{The html body data}

\item{template}{The html header/footer template}

\item{filename}{The name of the file to save the html. Default is a temp file}
}
\value{
The (local) url to the html file
}
\description{
Wrap html body in the template and save
}
