% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sotu}
\alias{sotu}
\title{Tokens from Bush and Obama's State of the Union addresses}
\format{sotu: A data.frame with tokens and a data.frame with meta data}
\usage{
data(sotu)
}
\description{
Tokens from Bush and Obama's State of the Union addresses
}
\keyword{datasets}
