% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sotu_data}
\alias{sotu_data}
\title{Tokens from Bush and Obamas State of the Union addresses}
\format{
sotu_data: A data.frame with tokens and a data.frame with meta data
}
\usage{
data(sotu_data)
}
\description{
Tokens from Bush and Obamas State of the Union addresses
}
\keyword{datasets}
