% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sotu_lda}
\alias{sotu_lda}
\title{Word assignments, docXtopic matrix and topicXword matrix of an LDA model of the SOTU data}
\format{
sotu_lda: Word assignments is a data.frame with document, lemma and topic columns. topic_word_mat and doc_topic_mat are matrices
}
\usage{
data(sotu_lda)
}
\description{
Word assignments, docXtopic matrix and topicXword matrix of an LDA model of the SOTU data
}
\keyword{datasets}
