% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{tag_attr}
\alias{tag_attr}
\title{create attribute string for html tags}
\usage{
tag_attr(...)
}
\arguments{
\item{...}{named arguments are used as attributes, with the name being the name of the attribute (e.g., class, style). All argument must be vectors of the same length, or lenght 1 (used as a constant). NA values can be used to skip an attribute. If all attributes are NA, an NA is returned}
}
\value{
a character vector with attribute strings. Designed to be usable as the attr_str in add_tag(). If ... is empty, NA is returned
}
\description{
create attribute string for html tags
}
\examples{
add_tag('TEXT', 'span')
add_tag('TEXT', 'span', tag_attr(class='CLASS'))
}
