% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_tokens.r
\name{topic_highlight_tokens}
\alias{topic_highlight_tokens}
\title{Highlight tokens per topic}
\usage{
topic_highlight_tokens(tokens, topic, alpha = 0.4, colors = NULL)
}
\arguments{
\item{tokens}{A character vector of tokens}

\item{topic}{A numeric vector with values representing topic indices.}

\item{alpha}{Optionally, the alpha (transparency) can be specified, with 0 being fully transparent and 1 being
fully colored. This can be a vector to specify a different alpha for each value.}

\item{colors}{A character vector with color names for unique values of the value argument. Has to be the same length
as unique(na.omit(topic))}
}
\value{
a character vector of color-tagged tokens
}
\description{
This is a convenience wrapper for tag_tokens() that can be used if tokens need to be colored per topic
}
