% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenvis.r
\name{topic_reader}
\alias{topic_reader}
\title{Convert tokens into full texts in an HTML file with color ramp highlighting}
\usage{
topic_reader(tokens, topic, alpha = 0.4, labels = levels(topic),
  meta = NULL, colors = NULL, doc_col = "doc_id", token_col = "token",
  filename = NULL, ...)
}
\arguments{
\item{tokens}{A data.frame with a column for document ids (doc_col)
and a column for tokens (token_col)}

\item{topic}{A numeric vector with values representing topic indices. Can also be a factor vector, in which case
the factor levels are automatically used as labels}

\item{alpha}{Optionally, the alpha (transparency) can be specified, with 0 being fully
transparent and 1 being fully colored. This can be a vector to specify a
different alpha for each value.}

\item{labels}{A character vector giving names to the unique values. If topic is a factor vector, the factor levels are
used.}

\item{meta}{A data.frame with a column for document_ids (doc_col). All other columns are added
to the reader as document meta}

\item{colors}{A character vector with color names for unique values of the topic argument. Has to be the same length
as unique(na.omit(topic))}

\item{doc_col}{The name of the document id column}

\item{token_col}{The name of the token column}

\item{filename}{Name of the output file. Default is temp file}

\item{...}{Additional formatting arguments passed to create_reader()}
}
\value{
The name of the file where the reader is saved. Can be opened conveniently from within R using browseUrl()
}
\description{
Convert tokens into full texts in an HTML file with color ramp highlighting
}
