% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_view.r
\name{view_reader}
\alias{view_reader}
\title{View a reader (HTML) in the rstudio viewer}
\usage{
view_reader(url)
}
\arguments{
\item{url}{An URL, created with *_reader}
}
\description{
View a reader (HTML) in the rstudio viewer
}
