% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_tokens.r
\name{category_highlight_tokens}
\alias{category_highlight_tokens}
\title{Highlight tokens per category}
\usage{
category_highlight_tokens(tokens, category, alpha = 0.4, colors = NULL)
}
\arguments{
\item{tokens}{A character vector of tokens}

\item{category}{A numeric vector with values representing category indices.}

\item{alpha}{Optionally, the alpha (transparency) can be specified, with 0 being fully transparent and 1 being
fully colored. This can be a vector to specify a different alpha for each value.}

\item{colors}{A character vector with color names for unique values of the value argument. Has to be the same length
as unique(na.omit(category))}
}
\value{
a character vector of color-tagged tokens
}
\description{
This is a convenience wrapper for tag_tokens() that can be used if tokens need to be colored per category
}
