% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenvis.r
\name{create_browser}
\alias{create_browser}
\title{Convert tokens into full texts in an HTML file}
\usage{
create_browser(tokens, meta = NULL, doc_col = "doc_id",
  token_col = "token", doc_nav = NULL, token_nav = NULL,
  filename = NULL, doc_width = 750, css_str = NULL, header = NULL,
  n = TRUE)
}
\arguments{
\item{tokens}{A data.frame with a column for document ids (doc_col)
and a column for tokens (token_col)}

\item{meta}{A data.frame with a column for document_ids (doc_col). All other columns are added
to the browser as document meta}

\item{doc_col}{The name of the document id column}

\item{token_col}{The name of the token column}

\item{doc_nav}{The name of a column (factor or character) in meta, used to create a navigation bar for selecting document groups.}

\item{token_nav}{Alternative to doc_nav, a column in the tokens.}

\item{filename}{Name of the output file. Default is temp file}

\item{doc_width}{The width of the document text field}

\item{css_str}{A character string, to be directly added to the css style header}

\item{header}{Optionally, specify the header}

\item{n}{If TRUE, report N in header}
}
\value{
The name of the file where the browser is saved. Can be opened conveniently from within R using browseUrl()
}
\description{
Convert tokens into full texts in an HTML file
}
\examples{
\dontrun{
url = create_browser(d$tokens, d$meta, token_col = 'token', header = 'Speeches')
view_browser(url)   ## view browser in the Viewer
browseURL(url)     ## view browser in default webbrowser
}
}
