% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.r
\name{html_template}
\alias{html_template}
\title{create the html template}
\usage{
html_template(template, doc_width = 750, css_str = NULL)
}
\arguments{
\item{template}{The name of the template to be used}

\item{doc_width}{The width of the document text field}

\item{css_str}{A character string, to be directly added to the css style header}
}
\value{
A list with the html header and footer
}
\description{
create the html template
}
